allow Integer itemValue for selectMultiMenu besides String only
Currently, only string values are allowed for selectMultiMenu items. A
typical use case for such a component is to have a description in the
label and an integer in the value. Therefore, I added a cast to Integer
for the itemValue and convert it to a String, so that the following
construction should output a list of itemValues instead of a list of
index values:

public class MyBean() {
  private final Integer key;
  private String value;
  ...
}

public List<MyBean> myList;
...
public String myValue;
...

<b:selectMultiMenu value="#{myValue}">
  <f:selectItems value="#{myList}" var="l" itemValue="#{l.key}"
itemLabel="#{l.value}" />
</b:selectMultiMenu>

One could argue that not only Integer, but also Float or Boolean can
become important - well, the solution is easy. But imho, Integer ist
much more important as I use this on database level as a primary key -
description pair, mostly the primary key is from a generated index (and
this is a number).