Fix expanded-panels inside tab
Further development showed up that a panel of an accordion inside a tab gets a  more depth client id. Now it takes the last part of the client id for comparison. Due to the fact that it is required that a panel is the direct child of an accordion it should be the correct id.